+++ tests/testthat/test-trajectory-select.R
@@ -1,4 +1,4 @@
-# Copyright (C) 2016-2017 Iñaki Ucar
+# Copyright (C) 2016-2018 Iñaki Ucar
 #
 # This file is part of simmer.
 #
@@ -54,6 +54,35 @@ test_that("core selection algorithms work: shortest-queue", {
   expect_equal(res_ordered$resource, c("r3", "r1", "r3", "r1", "r2", "r3"))
 })
 
+test_that("core selection algorithms work: shortest-queue-available", {
+  t0 <- trajectory() %>% seize("r1", 1)
+  t1 <- trajectory() %>% seize("r2", 1)
+
+  t2 <- trajectory() %>%
+    select(c("o1", "r1", "o2", "r2", "r3"), policy = "shortest-queue-available") %>%
+    seize_selected(1)
+
+  env <- simmer(verbose = TRUE) %>%
+    add_resource("o1", 0) %>%
+    add_resource("o2", 0) %>%
+    add_resource("r1", 2) %>%
+    add_resource("r2", 3) %>%
+    add_resource("r3", 1) %>%
+    add_generator("dummy0", t0, at(0)) %>%
+    add_generator("dummy1", t1, at(0, 0)) %>%
+    add_generator("dummy2", t2, at(seq(1, 6)))
+
+  expect_warning(env %>% run)
+
+  res <- get_mon_resources(env)
+  res_ordered <- res[order(res$time), ]
+  res_ordered <- res_ordered[4:9, ]
+
+  expect_equal(res_ordered$server, c(1, 2, 1, 2, 3, 1))
+  expect_equal(res_ordered$queue, c(0, 0, 1, 1, 0, 2))
+  expect_equal(res_ordered$resource, c("r3", "r1", "r3", "r1", "r2", "r3"))
+})
+
 test_that("core selection algorithms work: round-robin", {
   t0 <- trajectory() %>% seize("r1", 1)
   t1 <- trajectory() %>% seize("r2", 1)
@@ -81,15 +110,46 @@ test_that("core selection algorithms work: round-robin", {
   expect_equal(res_ordered$resource, c("r1", "r2", "r3", "r1", "r2", "r3"))
 })
 
+test_that("core selection algorithms work: round-robin-available", {
+  t0 <- trajectory() %>% seize("r1", 1)
+  t1 <- trajectory() %>% seize("r2", 1)
+
+  t2 <- trajectory() %>%
+    select(c("o1", "r1", "o2", "r2", "r3"), policy = "round-robin-available") %>%
+    seize_selected(1)
+
+  env <- simmer(verbose = TRUE) %>%
+    add_resource("o1", 0) %>%
+    add_resource("o2", 0) %>%
+    add_resource("r1", 2) %>%
+    add_resource("r2", 3) %>%
+    add_resource("r3", 1) %>%
+    add_generator("dummy0", t0, at(0)) %>%
+    add_generator("dummy1", t1, at(0, 0)) %>%
+    add_generator("dummy2", t2, at(seq(1, 6)))
+
+  expect_warning(env %>% run)
+
+  res <- get_mon_resources(env)
+  res_ordered <- res[order(res$time), ]
+  res_ordered <- res_ordered[4:9, ]
+
+  expect_equal(res_ordered$server, c(2, 3, 1, 2, 3, 1))
+  expect_equal(res_ordered$queue, c(0, 0, 0, 1, 1, 1))
+  expect_equal(res_ordered$resource, c("r1", "r2", "r3", "r1", "r2", "r3"))
+})
+
 test_that("core selection algorithms work: first-available", {
   t0 <- trajectory() %>% seize("r1", 1)
   t1 <- trajectory() %>% seize("r2", 1)
 
   t2 <- trajectory() %>%
-    select(c("r1", "r2", "r3"), policy = "first-available") %>%
+    select(c("o1", "r1", "o2", "r2", "r3"), policy = "first-available") %>%
     seize_selected(1)
 
   env <- simmer(ver = T) %>%
+    add_resource("o1", 0) %>%
+    add_resource("o2", 0) %>%
     add_resource("r1", 2, 1) %>%
     add_resource("r2", 3, 1) %>%
     add_resource("r3", 1, 1) %>%
@@ -128,6 +188,36 @@ test_that("core selection algorithms work: random", {
   expect_warning(res2 <- get_mon_resources(env %>% reset %>% run))
   expect_warning(res3 <- get_mon_resources(env %>% reset %>% run))
   expect_true(!all(res1 == res2) || !all(res1 == res3))
+})
+
+test_that("core '-available' algorithms fail if no resource is available", {
+  t <- function(policy) trajectory() %>%
+    select("r1", policy = policy)
+
+  expect_error(
+    simmer(verbose = TRUE) %>%
+      add_resource("r1", 0) %>%
+      add_generator("dummy", t("shortest-queue-available"), at(0)) %>%
+      run()
+  )
+  expect_error(
+    simmer(verbose = TRUE) %>%
+      add_resource("r1", 0) %>%
+      add_generator("dummy", t("round-robin-available"), at(0)) %>%
+      run()
+  )
+  expect_error(
+    simmer(verbose = TRUE) %>%
+      add_resource("r1", 0) %>%
+      add_generator("dummy", t("first-available"), at(0)) %>%
+      run()
+  )
+  expect_error(
+    simmer(verbose = TRUE) %>%
+      add_resource("r1", 0) %>%
+      add_generator("dummy", t("random-available"), at(0)) %>%
+      run()
+  )
 })
 
 test_that("custom selection algorithms work", {